Move workspace helper functions to it own files
