+++ src/commons/sagas/WorkspaceSaga/helpers/restoreExtraMethods.ts
@@ -1,0 +1,30 @@
+import { Context } from 'js-slang';
+import { call } from 'redux-saga/effects';
+
+import { getDifferenceInMethods, getRestoreExtraMethodsString } from '../../../utils/JsSlangHelper';
+import { EVAL_SILENT, WorkspaceLocation } from '../../../workspace/WorkspaceTypes';
+import { evalCode } from './evalCode';
+
+export function* restoreExtraMethods(
+  elevatedContext: Context,
+  context: Context,
+  execTime: number,
+  workspaceLocation: WorkspaceLocation,
+  unblockKey: string
+) {
+  const toUnblock = getDifferenceInMethods(elevatedContext, context);
+  const restorer = getRestoreExtraMethodsString(toUnblock, unblockKey);
+  const restorerFilePath = '/restorer.js';
+  const restorerFiles = {
+    [restorerFilePath]: restorer
+  };
+  yield call(
+    evalCode,
+    restorerFiles,
+    restorerFilePath,
+    elevatedContext,
+    execTime,
+    workspaceLocation,
+    EVAL_SILENT
+  );
+}
